added index settings stuff

+++ src/api/index.tsx
@@ -1,27 +1,27 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { amcatUserRoles, amcatIndexSchema } from "@/schemas";
-import { AmcatIndexName, AmcatUserRole } from "@/interfaces";
+import { AmcatIndexId, AmcatUserRole } from "@/interfaces";
 import { useHasGlobalRole } from "./userDetails";
 import { useAmcatConfig } from "./config";
 import { MiddlecatUser } from "middlecat-react";
 import { toast } from "sonner";
 
-export function useIndex(user?: MiddlecatUser, indexName?: AmcatIndexName) {
+export function useIndex(user?: MiddlecatUser, indexId?: AmcatIndexId) {
   return useQuery({
-    queryKey: ["index", user, indexName],
-    queryFn: async () => getIndex(user, indexName),
-    enabled: !!user && !!indexName,
+    queryKey: ["index", user, indexId],
+    queryFn: async () => getIndex(user, indexId),
+    enabled: !!user && !!indexId,
   });
 }
 
-export function useMyIndexrole(user?: MiddlecatUser, indexName?: AmcatIndexName) {
-  const { data: index } = useIndex(user, indexName);
+export function useMyIndexrole(user?: MiddlecatUser, indexId?: AmcatIndexId) {
+  const { data: index } = useIndex(user, indexId);
   return index?.user_role;
 }
 
-export function useHasIndexRole(user: MiddlecatUser, indexName: AmcatIndexName, role: AmcatUserRole) {
+export function useHasIndexRole(user: MiddlecatUser, indexId: AmcatIndexId, role: AmcatUserRole) {
   const { data: serverConfig } = useAmcatConfig();
-  const index_role = useMyIndexrole(user, indexName);
+  const index_role = useMyIndexrole(user, indexId);
   const is_global_admin = useHasGlobalRole(user, "ADMIN");
   if (is_global_admin) return true;
   if (serverConfig?.authorization === "no_auth") return true;
@@ -31,14 +31,14 @@ export function useHasIndexRole(user: MiddlecatUser, indexName: AmcatIndexName, 
   return actual_role_index >= required_role_index;
 }
 
-async function getIndex(user?: MiddlecatUser, indexName?: string) {
-  if (!user || !indexName) return undefined;
-  const res = await user.api.get(`/index/${indexName}`);
+async function getIndex(user?: MiddlecatUser, indexId?: string) {
+  if (!user || !indexId) return undefined;
+  const res = await user.api.get(`/index/${indexId}`);
   return amcatIndexSchema.parse(res.data);
 }
 
 interface MutateIndexParams {
-  id: AmcatIndexName;
+  id: AmcatIndexId;
   action: "create" | "delete" | "update";
   name?: string;
   description?: string;
@@ -46,7 +46,7 @@ interface MutateIndexParams {
   guest_role?: AmcatUserRole;
 }
 
-export function useMutateIndex(user: MiddlecatUser, indexName?: AmcatIndexName) {
+export function useMutateIndex(user: MiddlecatUser | undefined) {
   const queryClient = useQueryClient();
 
   return useMutation({
@@ -55,7 +55,8 @@ export function useMutateIndex(user: MiddlecatUser, indexName?: AmcatIndexName) 
       return mutateIndex(user, id, action, name, description, summary_field, guest_role);
     },
     onSuccess: (_, variables) => {
-      queryClient.invalidateQueries({ queryKey: ["index", user, indexName] });
+      const indexId = variables.id;
+      queryClient.invalidateQueries({ queryKey: ["index", user, indexId] });
       queryClient.invalidateQueries({ queryKey: ["indices", user] });
 
       if (variables.action === "create") toast.success(`Created index ${variables.id}`);
@@ -69,14 +70,15 @@ export function useMutateIndex(user: MiddlecatUser, indexName?: AmcatIndexName) 
 }
 
 export async function mutateIndex(
-  user: MiddlecatUser,
-  id: AmcatIndexName,
+  user: MiddlecatUser | undefined,
+  id: AmcatIndexId,
   action: "create" | "delete" | "update",
   name?: string,
   description?: string,
   summary_field?: string,
   guest_role?: AmcatUserRole,
 ) {
+  if (!user) throw new Error("Not logged in");
   if (action === "delete") {
     return await user.api.delete(`index/${id}`);
   } else if (action === "create") {
